add profile tag linking user to posts and likes

+++ app/Http/Controllers/PostController.php
@@ -14,7 +14,9 @@ class PostController extends Controller
     public function index(){
         //$posts = Post::get(); // return all posts as laravel collection
         // to reduce queries for likes for each post we are going to eager load it with the appropriate relationship
-        $posts = Post::with(['user','likes'])->paginate(20);
+        // "latest" = orderBy('created_at','desc')
+
+        $posts = Post::latest()->with(['user','likes'])->paginate(20);
         return view('posts.index', ['posts' => $posts]);
     }
     public function store(Request $request){
@@ -24,7 +26,8 @@ class PostController extends Controller
         $request->user()->posts()->create($request->only('body'));
         return back();
     }
-    public function destroy(Request $request,Post $post){
+    public function destroy(Post $post){
+        $this->authorize('delete', $post);
         $post->delete();
         return back();
     }
